make moving aware of sent mail (allowing for Fcc)
By telling notmuch to add Fcc: header:
(setq notmuch-fcc-dirs "sent +sent -inbox -unread -unprocessed")

When we send a message, a copy gets dropped in our 'sent' maildir. This works
fine, and then gets synchronized to IMAP server, etc, but! The problem is that
if we have sent to ourself, another copy appears in the inbox. If that happens,
notmuch gets a little screwy, as it has one reference for the message, but
multiple filenames for it.

For example:

If we search for "folder:inbox and not tag:inbox" we find this multi-message --
as the file appears in the inbox, but since the message already existed in the
database, no tag was added when the duplicate file appeared in the inbox. If
it's in the inbox but doesn't have the tag, then the previous logic said that we
must have removed the tag, which means we wanted to archive it. So then we would
move the message out of both the sent and inbox folder into the archive. This
seems... wrong.

Now, we first get rid of the duplicates in the inbox -- which we can find with
the "--duplicate=2" flag, and then do the rest.