improve reconnection algorithm
Now in addition to checking for the status Disconnected/UndefinedState, which
seem to not be a reliable indicator that the connection has dropped (in
particular, whenever going to sleep for a while, when waking up, we stay in the
"Connected" state even though we clearly aren't), we check if it's been more
than 13 minutes since we've last checked. This is a hack to detect when we've
been asleep, since otherwise we should be checking every 10 seconds. The time 13
is arbitrary -- could have been anything beyond a few minutes.